+++ src/main/java/br/com/guilhermeandrade/foods/food/FoodResponseDTO.java
@@ -1,6 +1,8 @@
 package br.com.guilhermeandrade.foods.food;
 
-public record FoodResponseDTO(Long id, String title, String image, Integer price) {
+import java.math.BigDecimal;
+
+public record FoodResponseDTO(Long id, String title, String image, BigDecimal price) {
     public FoodResponseDTO(Food food){
         this(food.getId(), food.getTitle(), food.getImage(), food.getPrice());
     }
